Consider the fact that there may be root bones.
Root bones have a parent index of UINT16_MAX (as far as I can tell). In this case, the parent is set to nil. Also, because I know I'll need it later, the model can return the root bones now; this is simply the set of bones with no parent.