Description of what shaders belong to what mesh group.
XNALara selects the shader and whether to draw with alpha or not based on the name of the mesh group. Okay…
The name of the mesh group has no relation at all to the shader or the draw mode; it seems to be assigned by fair dice roll. Not good…
The mapping between the mesh group and shader is entirely hardcoded. That's bad…
The mapping is not defined in any one place, but all over the RenderSceneTotal method, in lots of stunningly similar code, none of which has the slightest comment. We're into TheDailyWTF territory here.

I've decided to keep the mesh group names as they are (for now…), but store the mapping in a central location. Specifically, this file. That could have been code just as easily (no non-developer will ever change it anyway), but this way, I get the added benefit of having it all in a plist without having to type all the damn @s.