Further notes on how to understand XNALara
Of course I can't get around the hardcoding. There are tables, these tables are important, and at some place, I'll have to include them in my program, too. I'd just like it all to be a bit more organized, and if at all possible, have a system that is significantly easier to understand and maintain.