Transforms are now part of a uniform block, too.
The global uniforms are not a lot, but they're not nothing either, and if I never change the binding, only the contents of the buffer, I can save a lot of calls. And possibly confuse some readers, but I'll promise to write comments explaining the idea.

Note that OpenGL gives no guarantee that two shaders have to have the same layout for identical uniform blocks, if you don't use std140. I would be very surprised if there wasn't at least one graphics card (or driver) out to kill me, so here, where this buffer is needed all the time, I shall use std140, so that all shaders can share it.