Setting the locale to english, because numbers have a . as decimal separator in the ASCII format.
I can expect that the scanner will always use an english locale as long as it is always english. The minute my app becomes localized, though, the scanner will helpfully pick the locale closest to the user preferences, which may cause incorrect results, especially if that localization uses a , as decimal separator. So better set the locale to one I know works right from the start.