Changed how matrices are passed in; change projection when aspect ratio changes.
Now I'm just assuming that the bone matrices are always in global space. That means I need to have a fast matrix batch multiplication routine somewhere, and a place to use it, but that will come in time. As a result, the uniform block is now just one matrix, but that's still better than putting in a uniform because I can update it once, and it changes for absolutely all defined shaders.