Window system updated.
First of all, rendering now happens in its own window. It is possible to have several of these windows, but since they all have their own resources, not shared, I will probably not expose this feature in the user interface for now. The relationship between drawer, view and the newly added window controller has been clarified: The drawer always has a view available. It manages its own resources. The view informs the window controller when it is ready to do fun OpenGL stuff. At that point this creates the drawer.