Multiple windows per document!
Note: It's fun to see that this is possible, but I won't actually use
it. The reason is that each render window has its own textures, vertex
arrays, shaders and so on. As a result, it takes several seconds to
open each new window, and that's just annoying. This feature may come
back later, after I've made it so that the windows share textures and
the like. Ideally all windows, not just the ones from one document.