Display name now uses folder instead of file name.
Thanks to XNAaral for pointing this out! The file name is always one of a fairly limited set, and for items that use the generic item method, it is always generic_item. The name of the parent directory, on the other hand, can usually be relied on to be different.