Can now hide meshes.
As the old saying goes, everything can be solved with another layer of abstraction. In that manner, I now have a transformedMeshDrawer (not the best possible name), which draws a mesh as part of an item, taking into account all the data the item has about that mesh (in the MeshSettings object). The only relevant data is, of course, whether this mesh should be drawn at all or not, but that's no excuse for not using a proper architecture.