Generic_item now works as it should.
The original architecture could not work, because it relied on all mesh names being available when the ModelParams object is being created - even though the meshes need the model params to be created in the first place. Now, all the parsing happens on-demand and works nicely with the model loading scheme. Most methods don't need to access anything but their own mesh's name.