Corrected length reading.
So it's not really a pascal string at all. My impression is that this is a very simple coding scheme: Add bytes, shifted to the left, while the top bit is 1. I've actually written something vaguely like that some years ago, because it's simple and more efficient than e.g. what UTF8 does. Unlike UTF8, there's no real way to tell the start of a sequence, or the length of one without parsing all of it, but that's okay in this context.