Be a tiny bit more lenient when parsing generic item mesh names
The generic item format is simple enough; why can't people use it
properly? For example, I've found a mesh that defines camera targets
with no bones… possibly. Maybe that _fo is just tacked on there for no
good reason. The same mesh also uses a file name with an underscore,
which I think cannot be parsed deterministically anymore - unless the
file name is anything up to an underscore, and after that underscore
only numbers. Kind of like [^_\n]+(?:_[^0-9.]+)*. Hm, gotta think about
that.