Added Core Data infrastructure
This was a major undertaking, but worth it. Core Data allows me to save
and load scenes very easily, with full, automatic support for all the
stuff like undo, redo, updating the view when stuff changes - whatever
you want, it's there.

I will write importers and exporters for the XNALara Scene format.
However, I am not using that format as a default, because I have a lot
of interesting ideas, and I cannot really extend the old scene format
for them. The old scene format is also not all that great, really. It's
method for storing paths is kind of wonky, as is really all path
handling in XNALara.