Automatic redraw works again.
Since this isn't a game, the view only updates when I tell it to, which is good in one way (more efficient resource usage) but bad in another (you don't see changes right away, unless I tell the view to redraw). So this redraw has to be managed. The previous version did that explicitly, but it wasn't doing a very good job of it - for example, it supported only one scene drawer per scene, something I want to get away from in the long term. The new version uses key-value observing, which wasn't really meant for this kind of thing, but works well nonetheless.