Can change lights
This was a fairly big change, although most of the infrastructure was
already in place. Lights are now stored in their own GLLLights class,
which can write itself to a uniform buffer (very handy) and is
monitored all the time by the drawer, giving live updates, since
anything else would suck. Of course, there is full undo and such,
because why accept anything less?