Added alpha blending.
This was harder than I imagined, because as it turns out, the alpha test is gone with OpenGL 3.2 Core Profile. Not that bad an idea if you think about it; I can discard pixels much earlier in the shader because I know the final alpha value basically right at the start. Still, it causes a lot of debugging and such, adding yet another uniform buffer and all that fun.