Implemented Static TRL NextGen Shader
Not sure what the name is from; it's just DiffuseBump without bone weights, and in fact that's exactly what I'm using here. Unlike all that XNA/DirectX/HLSL stuff, there is no reason for me to write a new shader if I already have one that matches and all the inputs are named the same.