Show error message when trying to open new new-style generic item file.
I do want support for that file at some point, but XNAaral (the creator) doesn't like the idea of it being open source. I could try to reverse-engineer it, but I don't want to disrespect the wishes of someone who helped me a lot, so I won't. But that leaves the question: What do I do until then? My solution: Tell the user about the limitation, and recommend using an ASCII version instead.