Added OpenGL resource sharing.
Previously, each window had to load all textures again, even if the textures were already loaded in a different window. That wasn't nice, but it especially meant large amounts of redundant VRAM use when opening multiple render windows for one scene (which has always been possible, by the way). With resource sharing, this problem is gone. It required me to change the view to no longer use the default initializer, because I create my own OpenGL context now, and the resource manager is now a singleton. The scene drawer still stores it internally, because that was the quickest way to make the change.

If you want to try out multi-window support, simply add another RenderWindowController in the makeWindowControllers method of the document. It works perfectly, I just don't have a good UI for it yet.