Added a camera class.
This is still WIP; I'm thinking about how the camera should reference the bones. Specifically, is a camera target a property of a mesh, or of a bone? I want users to define their own, so it would make a lot of sense if a camera target could have several bones from different meshes associated with it.