New and improved lighting.
The lighting implementation of XNALara is very, very non-standard and very hard to understand. It is
also extremely limited. There is no way to set a color for specular highlights, for example. And did I
mention the oddities? Some, like the refusal to do proper bump mapping for diffuse components, I have
already mentioned. The shadowDepth for each light, instead of a standard ambient behavior; that is
also something I've never seen anywhere else, and it doesn't really help anyone either.

So now that is gone. Translating an XNALara Scene to the new format isn't difficult; translating back
is not possible without data loss. That's bad, but kind of the idea: I want my users to be able to
specify more data. Because I don't want to write a second XNALara that happens to be on the Mac; I want
to write an app that is better than XNALara, and being on the Mac is just the first step of that.