New feature: Manipulate render parameters of model.
All render parameters being set at export time is very limiting. What if they're wrong? Then they have to be changed, reexported, put in the scene again, posed again… that's annoying. So instead, you can now set the parameters, per mesh, at runtime. The settings are local to each item, so you can have two of the same item with different values if you want. Of course, it all updates live while you're sliding.

In terms of code changes: The render parameter buffer is now handled by the TransformedMeshDrawer instead of the MeshDrawer. There are also descriptions for the render parameters, and a shader now knows what parameters affect it even though they are not set by the generic_item or Item subclass in XNALara. The settings are stored in a RenderParamater object, an amazingly boring class. The RenderParameterDescription includes a human-readable title (localized, which got a new string file), description (again localized) and minimum and maximum values for the slider. The UI uses a view-based Table View to show all of this, linked to an NSArrayController.