Removed the status display key (which wasn't used anymore). Then refactored some code.
The status display was only ever used to make the screen less empty. With the render parameters, this is no longer an issue, so it could go. Then, in turn, the object controller could be bound to the represented object (instead of the view controller), which is better architecture all around. That meant changing a lot keys in the XIB, but the result is better than before.