Added image for camera settings
This was just a quick-and-dirty effort. What I don't like is how I have
to set the file as a template manually, but on the other hand, all code
regarding that works fine. It should also have no problem with HiDiPi
modes (or whatever they're called), seeing how it is a PDF file.