Proper Retina support.
The GLLView now requests a buffer at the optimal size. That means pixel-based APIs (like glViewport; in fact, right now, only glViewport) now no longer get the same measurements as the camera (which uses points), but a transformed version (in pixels). This is best set in the view so the scene drawer has no use for the overall size of the screen anymore.