Display name is now a property of the item.
The main goal for this was to allow it to easily find how many other items already have a given display name. All other solutions would have had problems. With the new solution, it is also possible (in the future) to change the display name of a model, because seriously, why not?