Bone transformations are now set in a uniform block and global, instead of local per mesh.
This was possible because I discovered that I can have 1024 matrices that way, which should be enough for anyone. The new technology requires that the ItemDrawer handles all transformations. The block, once set, is used for all meshes, and will be kept around unless any angle in the skeleton changes. That means slightly more efficient rendering, and less work for me, because the mesh drawer does not need to know about the bone objects anymore.