Load and save poses.
This triggered a complete rebuild of the entire bone system. Now, the bones update their matrices explicitly when a value changes. KVO is nice for some use cases, but it also makes debugging hard and causes new potential trouble. I've also changed the order in which the multiplications happen to match XNALara, for pose compatibility.