Textures are now stored as URLs. UV layer is ignored; getting a default texture from the bundle is implemented as fallback.
What does the UV Layer part in XNALara do? The shader doesn't need it; it knows what texture it is from its order in the file, and uses that semantic information to find the correct texture layer, no matter where it is. So that's out. Also, the idea of storing textures as names, relative to the model, is gone. That never helped anyone, and it makes it harder for new file formats.