Proper loading of diffuse OBJ files.
That required quite a number of changes. Some bug fixes in the OBJ loader and the related classes were the easiest. Adding shaders was necessary, making the model parameters less prominent (because OBJ uses MTL files with very different semantics, so it makes sense to hide the details), better support for color render parameters, shaders know their render parameter descriptions, default culling is specified by the mesh and so on.