Setting controller when the selection changes.
And just like that, I'm almost back to where I started this morning. But now with multiple selection! That was not entirely trivial. In my setup, each view has its own NSObjectController that manages the data. This will do the right thing if there is multiple selection, too, so all is awesome. But I cannot bind them directly to the selection of the general tree controller. They just won't get updated properly. So I have to set their content again (changing it to something else, here just nil, in between) whenever the selection changes.