Added multiple value editing for color parameters, too.
That was the point of the whole exercise: Change the colors of OBJ
files in bulk. That works now, but it is by far the slowest operation
in the whole app. Significantly slower than rendering really huge
files. I'll have to look into that, but I think for most situations,
this will be acceptable.