Added category on NSArray and NSDictionary that provide subscripting on Lion.
Subscripting consists of two parts: Compiler support, and the actual methods called by the compiler. These methods can be added to absolutely anything; including NSArray and so on, on pre-10.8 systems. I had feared that they would interfere with 10.8, but they don't, so that's nice. (I'm not sure whether my version is being called or an optimized one of the specific subclass, but that doesn't really matter either).