Split the scene drawer into a scene drawer and a view drawer.
The new scene drawer is shared by all objects in a document; it uses the shared context for what little loading and unloading it does, and it manages display updates, selection (in the future) and so on. Right now, it also does the actual draw calls, because it is the one managing all the tiny objects, although the data around it is handled by other parties. The view drawer does most of what the scene drawer did, but does not handle the changing models loaded and observing them for changes. It does handle the changing of the transformation (but only its own), and it does handle changes of the light (since the camera position is part of the light state. Actually, one can make a good argument for changing that, come to think of it…).