Shortened a lot of code by expressing things in a functional way, especially using map.
I've got a lot of cases where I got one array/dictionary/whatever, and I want to create another one using the objects from the first transformed in some way. This is exactly what map is for, and it used to be rather ugly before. There are also some cases of makeObjectsPerformSelector: and the like.