Changed method signature to make the static analyzer happy.
GLLURLFromString returns a URL that the caller has to release. Following standard conventions, it therefore needs either Create or Copy in its name. Apparently, the static analyzer checks and enforces this, so I changed this to make the warning go away (and because keeping with standard conventions is always a good idea).