Selection changes in the view now properly update the source list.
That was quite a fight, since expandItem: doesn't work unless that item's parent is already expanded (or, presumably, otherwise known to the source list), because the outline view does not know about a child item of something that's not been expanded yet until it was. And it can't select it, obviously. So now the source controllers know of their parent, and when selecting, they get opened in reverse order (oldest first).