Reduce cull face mode state changes.
This part has always bugged me: All the enables and disables and face mode settings. That's just ugly. But since there is an infrastructure to track active state now, I can simply reuse that. I don't think it will make sense to track more features through this, though, as texture sharing and the like are not very widespread.