Fix issue #23
Lots of small parts. For one, another variable ensures that selections don't get updated twice. The key part, though, is to ensure that the proper accessors for adding and removing get called, i.e. the bulk ones (which weren't implemented before, come to think of it). That minimizes the KVO notifications sent and thus the somewhat expensive operations done on each of them in the mesh view controller. It turns out that to ensure these bulk accessors get called, you have to call exactly the corresponding NSArray method. Otherwise, they may get called, but only with one element at a time, which doesn't help any.