Published version 0.1.1.
Also, new format for the bundle version. Now it's simply two integers. The first is the number of the public release (here 2 because the first was 0.1.0). The second is the number of the unofficial release since then, in this case 0. I reserve the right to use further subdivisions, but this should probably cover everything for now.