Models and items can now have parents (and children)
This also required the bones to have several items, since a child uses
its ancestor's bones (that sounds weird) if they already have one with
the same name. I'm not entirely certain that this is necessary for the
model, but we'll see. All of course in the service of implementing #7.