Fixes #7 by making skeleton drawer aware of the possibility of subitems.
And that's it for clothes support! In the future, it may also make sense to add a "separate" feature, to remove a child item from its parent but keeping all the existing parts, but that will be left to another time. Also, the UI needs improvements, but Item UI needs improvements anyway.