Fix enormous slowdown when changing from selecting an item to a single bone.
A similar issue had before as #25, but this one had a different root cause. In particular, the reason is that with all the combined bones and stuff, opening a single bone takes a lot longer than it used to. That made another issue very visible: Setting a single bone as the only selection caused first all bones to be selected, then all of them removed from the selection, then one added back in. That was kind of stupid. Now the selection is simply cleared, and then the single bone added.

The long time for finding children still needs to be addressed, though.