Find the parent bone controller faster.
This solves the slowdown issue. The problem was that loading the parent controller asked every single other bone controller for its children, which caused a lot of work behind the scenes. This is no longer the case; now it only asks all bone controllers for their bone, which is incredibly faster.