Allow saving of modified meshes.
First and most important part for solving #43. With this, lots of basic
mesh modding can now be done in real time right from GLLara, and saved
in formats that everyone can use.

Note that it may fail horribly if the original model was an OBJ file or
had no tangents. That needs to be fixed, somehow. I'm not entirely sure
how yet, but the simplest thing would be to disable exports for them.