Can now load meshes where some textures can't be found.
This is necessary for several XNALara items that assume a fixed folder structure and use other model's textures. It will still fail if a model does not have enough textures for its shader. Since it's not possible to retrieve the correct model programmatically, the default texture for the identifier is loaded (if one exists). Currently, the user is not informed of this substitution, so issue #49 can't be closed yet.