Fixed crash due to assertion error when loading models where some textures are missing.
The problem was that unload wasn't called in case of an error. I stand by the unload pattern, because you never know when an OpenGL context exists, but it has to be applied consistently. Right now, failure to load a drawer does not remove the item from the document. That probably needs to be fixed.