Fixes #57.
Apparently setting the primitive value for the item relationship in the ItemMesh screwed something up. As a result, the meshes were empty by deletion time, meaning they wouldn't delete their textures or render parameters (because they didn't have references to them anymore)… but the textures and render parameters still had dangling references to the meshes. That resulted in multiple validation errors, as Core Data puts it.

To make something like this easier to debug in the future, the errors in this case are now logged. Also, bones are now deleted when the last item referring to them is gone. I first assumed that this was the problem. It apparently wasn't, but it's still much cleaner like this.