Fixes #60.
This sets up a global notification for changed textures. All views observe this notification and redraw themselves when the texture has changed. This is not exactly the cleanest possible way to set this up, but it works okay. In the long term, the texture probably needs some way to inform its users of any changes (in particular renames), which could then be used for the redraw notification.