Trying to work around what looks like a race condition.
In some cases, not always reproducible, there's a crash in -[GLLDocumentWindowController dealloc] in the unbind for the meshViewController, which somehow accesses the (non-existent, hence the crash) shader property on a GLLItem. This move increases several internal time limits to avoid that condition, and adds an autoreleasepool, so this crash can be localized more easily.