Added automated rendering tests.
The tests currently have no expected value and no way of comparing them, because I didn't get around to doing that yet (and am not sure
whether that's a good idea with OpenGL at all). Still, they already helped me find a few issues, which is why so many shaders are modified.