Coalesce vertex arrays.
This ensures that any model has, for any given vertex format, exactly one vertex array. The mesh drawers only have indices into it. This reduces the count of calls to glBindVertexArray for one scene from 3319 to 9. No typo. The next step will be to coalesce the draw calls as well. There is a lot of potential for speedup here.