Fix bug with incorrect vertex indices in previous commit.
The vertex format depends on the vertex count, but this wasn't part of equals, and in one case where the vertex count was reset, the vertex format wasn't changed. As a result, some vertex drawers were created for the wrong vertex arrays values, which created visual artifacts. This is fixed now.