Localize MainMenu
As the first step. This required a few changes because the .strings
files Xcode generates are almost, but not quite, identical to the ones
used for the previous DL localization stuff (they would have been,
coming from the same Apple source tool, but Will Shipley thought it’s
nicer to have the actual text on the left side instead of just an ID).
This only requires the old files, some regexp magic, and then it all
works.

(Note: May require a clean to show up.)