Use separate shaders with and without blending.
Before, every shader could conditionally call discard, which makes it harder for the driver to optimize things. Now they are hidden behind a define which is conditionally enabled, which makes the solid case faster and doesn't hurt the alpha case. As part of this, new first lines are passed in by the runtime. This clashes with the #version tag, so that is now also passed in by the runtime instead of stored in the shader, so it can still be first.