Implement new KVO stuff for render parameters
This is the big one, and damn it, it actually works! Confuses the hell
out of me, honestly.

One thing: To make it all work properly, the mesh view controller now
caches the views for render parameters and textures. Seems to have no
ill effects, and improves performance. Setting the ambient and diffuse
colors on 1000 meshes in a level now works with almost no trouble.