Keep parameter buffer local and upload it as needed.
The parameter buffer is now recreated as an NSData object whenever its contents change; with the result uploaded to OpenGL as required. This allows using the buffer directly for comparisons for the sorting that reduces draw calls and finds runs. That should improve performance somewhat, since the whole O(n^2) stuff is now literally just one memcmp.