Fix depth test and clear color not working on OS X 10.11
For some reason I assumed the OpenGL context passed in here would be the current one by default. That was true on 10.10 and below, but not anymore on 10.11. By setting it, the state is correct and all is well with the world.

(so Apple did change something about OpenGL here? Nice and all, but I still don't see ARB_buffer_storage anywhere, and there's zero reason to leave this out. In some cases it may help a little, and in all others, just have it call through to glBufferData… I digress.)

Fascinating to see how little actually broke, all things considered. In the olden days, the screen would probably have stayed blank entirely.