Treat alpha and solid meshes the same
For a long time, alpha and solid meshes were separated at birth for
rendering. But with all the dynamic sorting we have now, this is no
longer necessary, as long as the sort algorithm ensures that all alpha
blended meshes get sorted behind all solid meshes. Ensuring that is
trivial. With this feature, we can easily treat both the same for much
of the loading pipeline.

This also allows, in theory, adding a checkbox to allow users to
override whether a mesh has alpha blending or not. Not sure that’s a
good idea, but it might be a fun one.