Use combined fragment shader for normal objects
The combined fragment shader is not yet used for OBJ, though some
provisions for that are in already. I think it would make a lot of
sense to just combine it all into one, but that would require some
interesting conversions - for example, bumpGloss should probably be
specularExponent everywhere, and bumpAmount should simply be turned
into the specularColor.