Upgrade settings to latest recommend by Xcode and fix warnings.
Xcode wants to add more warnings. Nice, except I am already compiling
with -wall, -wextra and -werror, so I’m fairly certain most of these
are duplicate. Except for two casts that it didn’t like. One of those
is harmless (in the view drawer), the other (in the mesh draw data) is
an actual bug - there’s no need to cast this to a pointer here. It is
necessary to cast it from unsigned to signed, which is not nice, but
not actually a big problem in practice for now.