Add support for render group 40, 41
Wow, it was not easy to get documentation on this; in the end I settled
for
http://junkymana.deviantart.com/art/XNALara-XPS-Render-Groups-543000260
which links to a source that does not have full data. There’s a number
of render groups in that list that is still missing; I started with
those for which I had a test model. Seems to work.

Also this proves that the new shader infrastructure does its job well.