Fix inconsistent state when loading something fails.
Subtle but annoying: It was possible for initialisation of the
GLLItemMeshState to fail partway through, e.g. if a shader couldn’t be
compiled. Then the object would be deallocated, including removing all
observers - some of which hadn’t been registered yet. That triggered
exceptions and left everything in a broken half-state. Fixing that
allowed fixing a lot of other issues as well.