Add models to document by dropping on "No selection" view
Previously adding models to a new document always involved clicking
some button or menu item and then selecting the model directly. That
was annoying, especially when trying out a lot of different new models
quickly to see what does and does not work about them. To fix this, it
is now possible to add models by simply dragging them on the view „no
selection“ view directly.

Further steps will include adding some text that tells people that this
is possible, and adding files by dragging them in the objects list on
the left.