New feature: Add image as plane
Not really sure it’s needed, but it was a fun exercise: Any image file
(any supported by ImageIO and also DDS) can now be added directly. It
will show up as a rectangle that can be positioned in the world. Could
be useful for various special effects. Probably less so for
backgrounds, those are still better added in Photoshop after the fact,
but anything where the image must be at a point in space should be
useful.

Examples may include gun shot flashes, smoke, glass shards, posters and
similar.