Slight improvements to optional parts handling
Correctly deals with parts with alternate meshes, i.e. the optional
part contains some meshes with a „+“, to be shown when the part is
logically visible, and some with a „-“, to be shown when the part is
logically invisible, to fill gaps that the other would leave.

Of course now it doesn’t work correctly with parts that are initially
entirely invisible, so that’s another thing to tackle… Also still no
useful display names or sub-menus (not that I’ve seen those in the wild
anyway).