Increase version number for release 0.2.8
I hadn’t actually planned to make a new one this quickly, but the bug
with the optional items, in one of the key new features, was annoying
enough that I felt it was worth it. And that gave me the opportunity to
address a few other places as well.