Completely redo vertex data handling
This is what the previous commits have been leading up to. Now vertex data is no longer tied to the old XNALara file format with some hacks for special cases. Instead it can come from any source, in any format, interleaved, separate or weird mixtures, and it can all be combined together on the fly anyway. This is a necessary prerequisite for glTF support. In my opinion it makes the code much clearer and easier to understand, but YMMV.

Specific changes:
- Split models simply copy all vertices and just use a subset of elements, because I didn't want to go through all that trouble for these tiny things.
- Exporting files without tangents (in particular some newer XPS files) as binary now works.