Moar Swift
Necessary precondition to changing VertexAttribSemantic to a struct that contains a VertexAttribSemanticType and an Int for a layer, which is necessary for making the shader description work properly with the texcoord1 thing.

This is probably not the best way to handle tex coord sets and will fail when we get to GLTF, but it'll work for now.