Initial basic support for 3D mice
Took quite a bit of refactoring in different places, including both the language and the logic of GLLView, and a custom NSWindow subclass.

Still needs setup UI, and the code for "rotating around camera" yields results that I strongly dislike (that's an old bug), so that's going to be fixed.