Fix detail bump maps
I'm not 100% sure about this because this way they smooth out the model a bit, but it is close to what old XNALara used to do. I've added the averaging before normalizing here because otherwise the results could get rather off, which is the main thing I was trying to avoid.