Fix Space Mouse support by rewriting it completely
This uses the official driver instead of the IOHIDInputManager, because that method doesn't work in Ventura anymore. I can still see the device in the IOHIDInputManager, but I can't open it to read devices because apparently it has been opened in exclusive mode.

Going via the official driver client API is a bit annoying, especially since I can't use the headers since it's supposed to build (as well as run!) without that driver installed. This also requires getting a hardened runtime entitlement to load code from elsewhere (specifically 3DConnexion).

This drops support for all 3D mouse type devices that aren't from 3Dconnexion, but since there don't seem to be any, that's probably okay.